Extend to split multiple ways.
Example only of splitting N ways, set up by the dests array.
Each saves to a file that is the "<array index>_filename" so
that each encoded stream can be saved somewhere unique.

JPEG seems to work OK with image_encode, but BMP stalls. This is
possibly not so surprising as it involves a conversion to RGB and
is not hardware accelerated.
Strong recommendation to only send frames to image_encode when
wanted. Don't try and use it as a video codec.